Refactor and use a model for password estimation.
